chore: support loader as string
