+++ packages/rspack/tests/configCases/loader/loader-string/index.js
@@ -1,0 +1,4 @@
+import { lib } from "./lib";
+it("loader-as-string", () => {
+	expect(lib).toEqual(43);
+});
