+++ packages/rspack/tests/configCases/loader/loader-string/lib.js
@@ -1,0 +1,1 @@
+export const lib = 42;
