+++ packages/rspack/tests/configCases/loader/loader-string/my-loader.js
@@ -1,0 +1,7 @@
+module.exports = function (context) {
+	return {
+		content: context.source.getCode().replace("42", "43"),
+		map: undefined,
+		meta: ""
+	};
+};
