+++ packages/rspack/tests/configCases/loader/loader-string/webpack.config.js
@@ -1,0 +1,18 @@
+/**
+ * @type {import('@rspack/core').RspackOptions}
+ */
+module.exports = {
+	context: __dirname,
+	module: {
+		rules: [
+			{
+				test: "lib.js",
+				uses: [
+					{
+						loader: "./my-loader.js"
+					}
+				]
+			}
+		]
+	}
+};
